+++ package.json
@@ -18,6 +18,7 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-redux": "^9.2.0",
+    "react-router-dom": "^7.6.0",
     "redux-persist": "^6.0.0"
   },
   "devDependencies": {
