+++ src/pages/Results.tsx
@@ -1,0 +1,8 @@
+export const Results = () => {
+  return (
+    <div>
+      <h1>Results</h1>
+      <p>This is the results page.</p>
+    </div>
+  );
+};
